+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add support for ubuntu >= 22.04 and debian >= 11
+
 ## 4.1.0 - *2022-07-20*
 
 - Remove duplicate pyenv init from profile script, fixes "pyenv: cannot rehash: <...>/shims isn't writable" for system install